Refactor RSA cryptor
