+++ AdyenCard/Validators/CardPublicKeyValidator.swift
@@ -11,7 +11,7 @@ internal struct CardPublicKeyValidator: Validator {
     
     /// :nodoc:
     public func isValid(_ string: String) -> Bool {
-        guard let range = string.range(of: #"([0-9]){5}\|([A-Z]|[0-9]){512}"#, options: .regularExpression) else {
+        guard let range = string.range(of: #"([0-9A-F]){5}\|([A-F]|[0-9]){512}"#, options: .regularExpression) else {
             return false
         }
         return range == (string.startIndex..<string.endIndex)
