+++ AdyenTests/Card Tests/Validators/RandomStringGenerator.swift
@@ -8,7 +8,7 @@ import Foundation
 
 enum RandomStringGenerator {
     static func generateDummyCardPublicKey() -> String {
-        return "\(generateRandomNumericString(length: 5))|\(generateRandomAlphaNumericString(length: 512))"
+        return "\(generateRandomHexadecimalString(length: 5))|\(generateRandomHexadecimalString(length: 512))"
     }
     
     static func generateRandomNumericString(length: Int) -> String {
@@ -16,8 +16,8 @@ enum RandomStringGenerator {
         return generateRandomString(length: length, characters: characters)
     }
     
-    static func generateRandomAlphaNumericString(length: Int) -> String {
-        let characters: [Character] = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z", "0", "1", "2", "3", "4", "5", "6", "7", "8", "9"]
+    static func generateRandomHexadecimalString(length: Int) -> String {
+        let characters: [Character] = ["A", "B", "C", "D", "E", "F", "0", "1", "2", "3", "4", "5", "6", "7", "8", "9"]
         return generateRandomString(length: length, characters: characters)
     }
